Updated QuizGame.py
Added game flow mechanics and logic, imports other classes. Will update again after Questions class is finalized.

NOTE: The questions are not supposed to be here in QuizGame class, but I put them here temporarily for my test run purposes. The questions are supposed to be in the Questions class. Please update the Questions class so I can also update how they get referenced as objects in the QuizGame class.